unit tests added, index.js code separated in 2 files
